+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.3.0
+
+- Change `global_get` method to store dict of raw value instead of obj.
+- ipx.inputs and ipx.outputs removed
+
+## 0.2.0
+
+- Change output and input from list to dict
 
 ## 0.1.1